(feat) change index/term from int to long

+++ src/main/java/raft/server/PendingProposalFutures.java
@@ -10,19 +10,19 @@ import java.util.concurrent.CompletableFuture;
  * Date: 18/6/3
  */
 class PendingProposalFutures {
-    private final TreeMap<Integer, CompletableFuture<ProposalResponse>> pendingProposal = new TreeMap<>();
+    private final TreeMap<Long, CompletableFuture<ProposalResponse>> pendingProposal = new TreeMap<>();
 
-    void addFuture(int lastIndex, CompletableFuture<ProposalResponse> responseFuture) {
+    void addFuture(long lastIndex, CompletableFuture<ProposalResponse> responseFuture) {
         assert lastIndex > 0;
         assert responseFuture != null;
 
         pendingProposal.put(lastIndex, responseFuture);
     }
 
-    void completeFutures(int commitIndex) {
-        NavigableMap<Integer, CompletableFuture<ProposalResponse>> fs = pendingProposal.headMap(commitIndex, true);
+    void completeFutures(long commitIndex) {
+        NavigableMap<Long, CompletableFuture<ProposalResponse>> fs = pendingProposal.headMap(commitIndex, true);
 
-        for (Map.Entry<Integer, CompletableFuture<ProposalResponse>> entry : fs.entrySet()) {
+        for (Map.Entry<Long, CompletableFuture<ProposalResponse>> entry : fs.entrySet()) {
             CompletableFuture<ProposalResponse> future = entry.getValue();
             future.complete(ProposalResponse.success());
             pendingProposal.remove(entry.getKey());
@@ -30,7 +30,7 @@ class PendingProposalFutures {
     }
 
     void failedAllFutures() {
-        for (Map.Entry<Integer, CompletableFuture<ProposalResponse>> entry : pendingProposal.entrySet()) {
+        for (Map.Entry<Long, CompletableFuture<ProposalResponse>> entry : pendingProposal.entrySet()) {
             CompletableFuture<ProposalResponse> future = entry.getValue();
             future.complete(ProposalResponse.error(ErrorMsg.NOT_LEADER));
             pendingProposal.remove(entry.getKey());
